# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Proyectos",
        "contact": {
            "email": "andy.ancajima@mimco.com.pe"
        },
        "version": "1"
    },
    "paths": {
        "/GestionProgramas/public/index.php/regi_usua": {
            "post": {
                "tags": [
                    "Gestion Agrupador"
                ],
                "summary": "permite el registro de un agrupador",
                "operationId": "App\\Http\\Controllers\\AgrupadorController::regi_agru",
                "parameters": [
                    {
                        "name": "varCodiAgru",
                        "in": "path",
                        "description": "Codigo para el agrupador",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "DESPPRUEBA"
                    },
                    {
                        "name": "varDescAgru",
                        "in": "path",
                        "description": "Descripcion de agrupador",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "GALVANIZADO EPOXICO"
                    },
                    {
                        "name": "acti_usua",
                        "in": "path",
                        "description": "usuario quien registra",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "usuario"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "varCodiAgru": {
                                        "type": "string"
                                    },
                                    "varDescAgru": {
                                        "type": "string"
                                    },
                                    "codi_usua": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "varCodiAgru": "KIRLM",
                                    "varEstaUsua": "ACT",
                                    "codi_usua": "usuario"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro Satisfactorio"
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    },
                    "408": {
                        "description": "Codigo de agrupador ya se encuentra registrado."
                    }
                }
            }
        },
...
        "/GestionProyectos/public/index.php/vali_agru": {
            "post": {
                "tags": [
                    "Gestion Agrupador"
                ],
                "summary": "obtiene datos del agrupador con su numero de id Agrupado",
                "operationId": "App\\Http\\Controllers\\AgrupadorController::vali_agru",
                "parameters": [
                    {
                        "name": "intIdAgru",
                        "in": "path",
                        "description": "documento de identidad",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdAgru": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdAgru": "5"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sin Mensaje"
                    },
                    "407": {
                        "description": "El Documento de identidad ingresado no se encuentra registrado."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
        "/GestionProyectos/public/index.php/actu_agru": {
            "post": {
                "tags": [
                    "Gestion Agrupador"
                ],
                "summary": "Permite actualizar el agrupador",
                "operationId": "App\\Http\\Controllers\\AgrupadorController::actu_agru",
                "parameters": [
                    {
                        "name": "varCodiAgru",
                        "in": "path",
                        "description": "Codigo para el agrupador",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "KIRLMPRUEBA"
                    },
                    {
                        "name": "varDescAgru",
                        "in": "path",
                        "description": "Descripcion de agrupador",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Habilitado"
                    },
                    {
                        "name": "codi_usua",
                        "in": "path",
                        "description": "usuario quien registra",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "usuario"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "varCodiAgru": {
                                        "type": "string"
                                    },
                                    "varDescAgru": {
                                        "type": "string"
                                    },
                                    "codi_usua": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "varCodiAgru": "KIRLMPRUEBA",
                                    "varDescAgru": "Habilitado",
                                    "codi_usua": "usuario"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizacion Satisfactorio"
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    },
                    "408": {
                        "description": "Codigo de agrupador ya se encuentra registrado."
                    }
                }
            }
        },
        "/GestionProyectos/public/index.php/list_etap_acue_agru": {
            "post": {
                "tags": [
                    "Gestion Agrupador"
                ],
                "summary": "Listar Etapa de acuerdo al agrupador",
                "operationId": "App\\Http\\Controllers\\AgrupadorController::list_etap_acue_agru",
                "parameters": [
                    {
                        "name": "intIdAgru",
                        "in": "path",
                        "description": "documento de identidad",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdAgru": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdAgru": "1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "muestra una lista deacuerdo a la etapa"
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
...
    }
}